requirements.txt
discord.py
python-dotenv
gspread
oauth2client
